MAINT: Fixed usage statistics export
